requirements.txt
apache-airflow-providers-apache-beam==4.3.0
apache-airflow-providers-celery==2.1.0
apache-airflow-providers-google==9.0.0
commonmark==0.9.1
flask-openid==1.3.0
iso8601==1.0.2
openapi-spec-validator==0.3.1
pip-chill==1.0.1
pytest==7.3.1
swagger-ui-bundle==0.0.9
